+++ libs/query/src/index.ts
@@ -1,4 +1,5 @@
 export * from './lib/auth';
+export * from './lib/class-validator';
 export * from './lib/directives';
 export * from './lib/entity';
 export * from './lib/gql';