+++ libs/query/src/lib/class-validator/class-validator.types.ts
@@ -1,0 +1,7 @@
+import { HttpStatusCode } from '../request';
+
+export interface ClassValidatorError {
+  statusCode: HttpStatusCode;
+  message: string[];
+  error: string;
+}